% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B.utils.R
\name{B.color}
\alias{B.color}
\title{Returns color vector}
\usage{
B.color(color = "darkblue1")
}
\arguments{
\item{color}{It can be the color name in which case the function will only
return that color or a number in which case the function will return a
vector of colors with the parameter's length. Defaults to "darkblue"}
}
\description{
This function returns one color or a vector of colors, similar to the
rainbow() function. The only difference is it uses the company color
palette
}
\examples{
B.color(10)
}

